// src/generated/models/BuildSetting.ts
export interface BuildSetting {
    /**
     * 
     * @type {number}
     * @memberof BuildSetting
     */
    readonly id: number;
    /**
     * 
     * @type {Date}
     * @memberof BuildSetting
     */
    readonly createdAt: Date;
    /**
     * 
     * @type {BuildSettingCreatedBy}
     * @memberof BuildSetting
     */
    createdBy: BuildSettingCreatedBy;
    /**
     * 
     * @type {Date}
     * @memberof BuildSetting
     */
    readonly updatedAt: Date;
    /**
     * 
     * @type {BuildSettingCreatedBy}
     * @memberof BuildSetting
     */
    updatedBy: BuildSettingCreatedBy;
    /**
     * 
     * @type {string}
     * @memberof BuildSetting
     */
    repository: string;
    /**
     * 
     * @type {string}
     * @memberof BuildSetting
     */
    dockerfilePath?: string;
    /**
     * 
     * @type {number}
     * @memberof BuildSetting
     */
    cpuLimit?: number;
    /**
     * 
     * @type {number}
     * @memberof BuildSetting
     */
    memoryLimit?: number;
    /**
     * 
     * @type {number}
     * @memberof BuildSetting
     */
    cpuRequest?: number;
    /**
     * 
     * @type {number}
     * @memberof BuildSetting
     */
    memoryRequest?: number;
    /**
     * 
     * @type {number}
     * @memberof BuildSetting
     */
    activeDeadline?: number;
}